fix: cast points_total back to bigint in classement_saison view
sum(bigint) returns numeric in Postgres, but the existing view column
was bigint (from the original sum(int) definition), so CREATE OR
REPLACE VIEW rejected the implicit type change. Explicit cast keeps
it bigint, matching what admin's Supabase project already has.

+++ supabase/schema.sql
@@ -150,7 +150,7 @@ create or replace view classement_saison as
 select
   joueur_id,
   pseudo,
-  sum(points_soiree) as points_total
+  sum(points_soiree)::bigint as points_total
 from classement_soiree
 group by joueur_id, pseudo
 order by points_total desc;
@@ -241,7 +241,7 @@ create or replace view classement_saison as
 select
   joueur_id,
   pseudo,
-  sum(points_soiree) as points_total
+  sum(points_soiree)::bigint as points_total
 from classement_soiree
 group by joueur_id, pseudo
 order by points_total desc;
